+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/SKDefaultShopObjectTypes.java
@@ -7,6 +7,7 @@ import com.nisovin.shopkeepers.SKShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.shopobjects.DefaultShopObjectTypes;
 import com.nisovin.shopkeepers.shopobjects.block.base.BaseBlockShops;
 import com.nisovin.shopkeepers.shopobjects.citizens.SKCitizensShopObjectType;
+import com.nisovin.shopkeepers.shopobjects.endcrystal.SKEndCrystalShopObjectType;
 import com.nisovin.shopkeepers.shopobjects.entity.base.BaseEntityShops;
 import com.nisovin.shopkeepers.shopobjects.living.SKLivingShopObjectTypes;
 import com.nisovin.shopkeepers.shopobjects.sign.SKHangingSignShopObjectType;
@@ -18,6 +19,7 @@ public final class SKDefaultShopObjectTypes implements DefaultShopObjectTypes {
 
 	private final SKShopkeepersPlugin plugin;
 
+	private final SKEndCrystalShopObjectType endCrystalShopObjectType;
 	private final SKSignShopObjectType signShopObjectType;
 	private final SKHangingSignShopObjectType hangingSignShopObjectType;
 
@@ -27,18 +29,21 @@ public final class SKDefaultShopObjectTypes implements DefaultShopObjectTypes {
 			BaseEntityShops entityShops
 	) {
 		this.plugin = plugin;
+		this.endCrystalShopObjectType = new SKEndCrystalShopObjectType(entityShops);
 		this.signShopObjectType = new SKSignShopObjectType(blockShops);
 		this.hangingSignShopObjectType = new SKHangingSignShopObjectType(blockShops);
 	}
 
 	public void onRegisterDefaults() {
 		this.getLivingShopObjectTypes().onRegisterDefaults();
+		endCrystalShopObjectType.registerPermission();
 	}
 
 	@Override
 	public List<? extends AbstractShopObjectType<?>> getAll() {
 		List<AbstractShopObjectType<?>> shopObjectTypes = new ArrayList<>();
 		shopObjectTypes.addAll(this.getLivingShopObjectTypes().getAll());
+		shopObjectTypes.add(this.getEndCrystalShopObjectType());
 		shopObjectTypes.add(this.getSignShopObjectType());
 		shopObjectTypes.add(this.getHangingSignShopObjectType());
 		shopObjectTypes.add(this.getCitizensShopObjectType());
@@ -48,6 +53,11 @@ public final class SKDefaultShopObjectTypes implements DefaultShopObjectTypes {
 	@Override
 	public SKLivingShopObjectTypes getLivingShopObjectTypes() {
 		return plugin.getLivingShops().getLivingShopObjectTypes();
+	}
+
+	@Override
+	public SKEndCrystalShopObjectType getEndCrystalShopObjectType() {
+		return endCrystalShopObjectType;
 	}
 
 	@Override
@@ -73,6 +83,10 @@ public final class SKDefaultShopObjectTypes implements DefaultShopObjectTypes {
 
 	public static SKLivingShopObjectTypes LIVING() {
 		return getInstance().getLivingShopObjectTypes();
+	}
+
+	public static SKEndCrystalShopObjectType END_CRYSTAL() {
+		return getInstance().getEndCrystalShopObjectType();
 	}
 
 	public static SKSignShopObjectType SIGN() {